EX-168: fix

+++ core/main.go
@@ -165,7 +165,7 @@ func (ext *Extender) Run() {
 		ext.handleAddressesFromResponses(blockResponse, eventsResponse)
 		ext.handleBlockResponse(blockResponse)
 
-		if height%ext.env.RewardAggregateEveryBlocksCount == 0 {
+		if height%uint64(ext.env.RewardAggregateEveryBlocksCount) == 0 {
 			go ext.eventService.AggregateRewards(ext.env.RewardAggregateTimeInterval, height)
 		}
 		go ext.handleEventResponse(height, eventsResponse)
